refactor: extract live room chat reactions
